requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
python-multipart>=0.0.9
google-generativeai>=0.5.0
sentence-transformers>=3.0.0
chromadb>=0.5.0
pymupdf>=1.24.0
python-dotenv>=1.0.0
pydantic>=2.7.0
torch
protobuf